+++ app/models/submission.rb
@@ -9,13 +9,14 @@ class Submission < ApplicationRecord
 
   validates_uniqueness_of :fund, scope: :organisation
   validate :activity_must_be_a_fund
+  validates :deadline, date_not_in_past: true, date_within_boundaries: true
 
-  enum state: [:inactive]
+  enum state: [:inactive, :active]
 
   def activity_must_be_a_fund
     return unless fund.present?
     unless fund.fund?
-      errors.add(:fund, I18n.t("errors.models.submission.attributes.fund"))
+      errors.add(:fund, I18n.t("activerecord.errors.models.submission.attributes.fund.level"))
     end
   end
 end