+++ spec/models/submission_spec.rb
@@ -15,7 +15,7 @@ RSpec.describe Submission, type: :model do
     programme = create(:programme_activity)
     submission = build(:submission, fund: programme)
     expect(submission).not_to be_valid
-    expect(submission.errors[:fund]).to include I18n.t("errors.models.submission.attributes.fund")
+    expect(submission.errors[:fund]).to include I18n.t("activerecord.errors.models.submission.attributes.fund.level")
   end
 
   it "does not allow more than one Submission for the same Fund and Organisation combination" do
@@ -25,5 +25,10 @@ RSpec.describe Submission, type: :model do
 
     new_submission = build(:submission, organisation: organisation, fund: fund)
     expect(new_submission).not_to be_valid
+  end
+
+  it "does not allow a Deadline which is in the past" do
+    submission = build(:submission, deadline: Date.yesterday)
+    expect(submission).not_to be_valid
   end
 end